Ajout d'une route pour générer une facture, qui gère toute la logique de facturation : fonctionnelle

+++ src/database/Cart.ts
@@ -10,10 +10,10 @@ export const Cart = sequelize.define("Cart", {
         defaultValue: false,
     },
     totalPrice: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.DECIMAL(10, 2), // Précision pour des valeurs monétaires
         allowNull: false,
-        defaultValue: 0,
-    }
+        defaultValue: 0.00,
+      }
 })
